+++ src/mat3.rs
@@ -473,12 +473,18 @@ impl Mat3 {
         self.mul_vec3(other.into()).into()
     }
 
+    #[deprecated(since = "0.18.0", note = "please use `as_dmat3()` instead")]
     #[inline(always)]
     pub fn as_f64(&self) -> DMat3 {
+        self.as_dmat3()
+    }
+
+    #[inline(always)]
+    pub fn as_dmat3(&self) -> DMat3 {
         DMat3::from_cols(
-            self.x_axis.as_f64(),
-            self.y_axis.as_f64(),
-            self.z_axis.as_f64(),
+            self.x_axis.as_dvec3(),
+            self.y_axis.as_dvec3(),
+            self.z_axis.as_dvec3(),
         )
     }
 }
@@ -508,12 +514,18 @@ impl Mat3A {
         Vec3A(self.0.mul_vector(other.0))
     }
 
+    #[deprecated(since = "0.18.0", note = "please use `as_dmat3()` instead")]
     #[inline(always)]
     pub fn as_f64(&self) -> DMat3 {
+        self.as_dmat3()
+    }
+
+    #[inline(always)]
+    pub fn as_dmat3(&self) -> DMat3 {
         DMat3::from_cols(
-            self.x_axis.as_f64(),
-            self.y_axis.as_f64(),
-            self.z_axis.as_f64(),
+            self.x_axis.as_dvec3(),
+            self.y_axis.as_dvec3(),
+            self.z_axis.as_dvec3(),
         )
     }
 }
@@ -547,12 +559,18 @@ define_mat3_struct!(DMat3, InnerF64);
 impl DMat3 {
     impl_mat3_methods!(f64, DVec3, DVec3, DVec2, DQuat, DMat2, DMat4, InnerF64);
 
+    #[deprecated(since = "0.18.0", note = "please use `as_mat3()` instead")]
     #[inline(always)]
     pub fn as_f32(&self) -> Mat3 {
+        self.as_mat3()
+    }
+
+    #[inline(always)]
+    pub fn as_mat3(&self) -> Mat3 {
         Mat3::from_cols(
-            self.x_axis.as_f32(),
-            self.y_axis.as_f32(),
-            self.z_axis.as_f32(),
+            self.x_axis.as_vec3(),
+            self.y_axis.as_vec3(),
+            self.z_axis.as_vec3(),
         )
     }
 }