+++ src/mat4.rs
@@ -790,13 +790,19 @@ impl Mat4 {
         Vec3A(self.0.transform_float4_as_vector3(other.0.into()).into())
     }
 
+    #[deprecated(since = "0.18.0", note = "please use `as_dmat4()` instead")]
     #[inline(always)]
     pub fn as_f64(&self) -> DMat4 {
+        self.as_dmat4()
+    }
+
+    #[inline(always)]
+    pub fn as_dmat4(&self) -> DMat4 {
         DMat4::from_cols(
-            self.x_axis.as_f64(),
-            self.y_axis.as_f64(),
-            self.z_axis.as_f64(),
-            self.w_axis.as_f64(),
+            self.x_axis.as_dvec4(),
+            self.y_axis.as_dvec4(),
+            self.z_axis.as_dvec4(),
+            self.w_axis.as_dvec4(),
         )
     }
 }
@@ -841,13 +847,19 @@ pub struct DMat4(pub(crate) InnerF64);
 impl DMat4 {
     impl_mat4_methods!(f64, DVec4, DVec3, DMat3, DQuat, InnerF64);
 
+    #[deprecated(since = "0.18.0", note = "please use `as_mat4()` instead")]
     #[inline(always)]
     pub fn as_f32(&self) -> Mat4 {
+        self.as_mat4()
+    }
+
+    #[inline(always)]
+    pub fn as_mat4(&self) -> Mat4 {
         Mat4::from_cols(
-            self.x_axis.as_f32(),
-            self.y_axis.as_f32(),
-            self.z_axis.as_f32(),
-            self.w_axis.as_f32(),
+            self.x_axis.as_vec4(),
+            self.y_axis.as_vec4(),
+            self.z_axis.as_vec4(),
+            self.w_axis.as_vec4(),
         )
     }
 }